requirements.txt
opencv-python
Pillow
numpy
# optional for drag-and-drop. On some systems install manually if fails:
TkinterDnD2
